requirements.txt
requests
flatlib
pyswisseph
google-generativeai
python-dotenv
python-multipart